requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.133.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
-e git+https://github.com/eduardoarsocca/integracao_natura.git@0cf2be99bd22182a830a10f99da4cbb590d31206#egg=integracao
integracao_redcap_polotrial==0.1.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycap==2.7.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
semantic-version==2.10.0
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
tornado==6.5.2
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
